Add limit to objdump's lrucache
