+++ backend/decompme/settings.py
@@ -36,6 +36,7 @@ env = environ.Env(
     COMPILER_BASE_PATH=(str, BASE_DIR / "compilers"),
     LIBRARY_BASE_PATH=(str, BASE_DIR / "libraries"),
     COMPILATION_CACHE_SIZE=(int, 100),
+    OBJDUMP_CACHE_SIZE=(int, 100),
     WINEPREFIX=(str, "/tmp/wine"),
     COMPILATION_TIMEOUT_SECONDS=(int, 10),
     ASSEMBLY_TIMEOUT_SECONDS=(int, 3),
@@ -229,6 +230,7 @@ GITHUB_CLIENT_ID = env("GITHUB_CLIENT_ID", str)
 GITHUB_CLIENT_SECRET = env("GITHUB_CLIENT_SECRET", str)
 
 COMPILATION_CACHE_SIZE = env("COMPILATION_CACHE_SIZE", int)
+OBJDUMP_CACHE_SIZE = env("OBJDUMP_CACHE_SIZE", int)
 
 WINEPREFIX = Path(env("WINEPREFIX"))
